requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
openai
pytest
SQLAlchemy>=2.0
psycopg2-binary
alembic
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-multipart